Move device JSON files to devices folder.
