+++ src/devices.h
@@ -1,31 +1,31 @@
-#include "HHCCJCY01HHCC_json.h"
-#include "VEGTRUG_json.h"
-#include "LYWSD02_json.h"
-#include "LYWSDCGQ_json.h"
-#include "CGP1W_json.h"
-#include "CGG1_json.h"
-#include "CGD1_json.h"
-#include "CGDK2_json.h"
-#include "CGH1_json.h"
-#include "JQJCY01YM_json.h"
-#include "IBS_TH1_json.h"
-#include "IBS_TH2_json.h"
-#include "IBS_X4S_json.h"
-#include "Miband_json.h"
-#include "XMTZC04HM_json.h"
-#include "XMTZC05HM_json.h"
-#include "TPMS_json.h"
-#include "LYWSD03MMC_ATC_json.h"
-#include "CGPR1_json.h"
-#include "iBeacon_json.h"
-#include "WS02_json.h"
-#include "H5075_json.h"
-#include "H5072_json.h"
-#include "LYWSD03MMC_PVVX_json.h"
-#include "MUE4094RT_json.h"
-#include "Mokobeacon_json.h"
-#include "MBXPRO_json.h"
-#include "iNode_json.h"
+#include "devices/HHCCJCY01HHCC_json.h"
+#include "devices/VEGTRUG_json.h"
+#include "devices/LYWSD02_json.h"
+#include "devices/LYWSDCGQ_json.h"
+#include "devices/CGP1W_json.h"
+#include "devices/CGG1_json.h"
+#include "devices/CGD1_json.h"
+#include "devices/CGDK2_json.h"
+#include "devices/CGH1_json.h"
+#include "devices/JQJCY01YM_json.h"
+#include "devices/IBS_TH1_json.h"
+#include "devices/IBS_TH2_json.h"
+#include "devices/IBS_X4S_json.h"
+#include "devices/Miband_json.h"
+#include "devices/XMTZC04HM_json.h"
+#include "devices/XMTZC05HM_json.h"
+#include "devices/TPMS_json.h"
+#include "devices/LYWSD03MMC_ATC_json.h"
+#include "devices/CGPR1_json.h"
+#include "devices/iBeacon_json.h"
+#include "devices/WS02_json.h"
+#include "devices/H5075_json.h"
+#include "devices/H5072_json.h"
+#include "devices/LYWSD03MMC_PVVX_json.h"
+#include "devices/MUE4094RT_json.h"
+#include "devices/Mokobeacon_json.h"
+#include "devices/MBXPRO_json.h"
+#include "devices/iNode_json.h"
 
 const char* _devices[] = {
 _HHCCJCY01HHCC_json,
