+++ src/devices/CGD1_json.h
