+++ src/devices/CGDK2_json.h
