+++ src/devices/CGG1_json.h
