+++ src/devices/CGH1_json.h
