+++ src/devices/CGP1W_json.h
