+++ src/devices/CGPR1_json.h
