+++ src/devices/H5072_json.h
