+++ src/devices/H5075_json.h
