+++ src/devices/HHCCJCY01HHCC_json.h
