+++ src/devices/LYWSD02_json.h
