+++ src/devices/LYWSD03MMC_PVVX_json.h
