+++ src/devices/LYWSDCGQ_json.h
