+++ src/devices/VEGTRUG_json.h
