+++ src/devices/WS02_json.h
